renamed _main to main cause gitlens is not working with _main
